busybox: crude reproducible build patches
These are patches to busybox to enable reproducible builds: aka, to have the
busybox executable come out bit-for-bit identical between the official openwrt
images and any third-party attempts to reproduce those images. In the case of
busybox, all that is needed to override the build-time timestamp.

The "111-source-timestamp.patch" could go upstream, but it would probably be
better to integrate with busybox's kbuild system to save the timestamp in a
header file (as is done now). SOURCE_DATE_EPOCH gets passed to the Configure
step as well, so this should work fine.

The busybox/Makefile patch proposes a new PKG_TIMESTAMP_EPOCH field (numerical
UNIX timestamp). This is thought to be necessary because the source is
distributed as a tarball and thus doesn't have version control info to derive a
SOURCE_DATE_EPOCH. It could alternatively be a string ("PKG_DATE"?) which would
be parsed by `date` to get a unix timestamp. This would be far more readable.