include/image.mk: export SOURCE_DATE_EPOCH
The utility of this patch is to achieve reproducible image builds.

The patch converts the top-level $TIMESTAMP into a numerical UNIX epoch
timestamp, and exports it in the SOURCE_DATE_EPOCH environment variable forthe
image creation process. It doesn't export this variable earlier, because
packages (.ipkg) or other build steps might be able to derive a
SOURCE_DATE_EPOCH from more-local version control metadata. In the case of
image artifacts, the timestamp (eg, on uImage generated image files, which
sometimes include the kernel et al) should be coming from the OpenWRT repo
timestamp.