tools/squashfs4: mksquashfs support for SOURCE_DATE_EPOCH
This is a local patch to give the mksquashfs tool support for the
SOURCE_DATE_EPOCH environment variable. If this variable is set to a valid UNIX
timestamp, both the generated squashfs image metadata, and the modification
timestamps of any special ("pseudo") files *within* the filesystem itself will
get that timestamp.

This is to support reproducible image builds.