kernel: set kernel build timestamp to top-level value
This sets the kernel's internal timestamp metadata to be the same as the entire
openwrt build process (which, with an earlier commit, is set to the most recent
git/subversion commit time).

The reason the Kernel's version control timestamp can't be used (instead of
OpenWRT's) is that this information is not contained in tarball archives.